Added move shapes to other layers
Shapes are moved by cuting them, changing the current location and pasting them again.
Currently it is a bit bugged, as it always pastes the shapes to the map layer. The workaround (which, lets be honest, will probably be permanent) is to move them to the right layer after pasting.
Also added delete the currently selected shapes